+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 *** 0.1.1
 
 - Changed default filter bindings changed to make projectile not a hard dependency
+- added `frog-jump-buffer-include-current-buffer` option